Fixed bug with p-value lower than precision of a double
Added a minimum value of 10^-16 for p-values

+++ src/utils.h
@@ -19,7 +19,7 @@ struct Locus {
     std::vector<std::string> coverage;
 };
 
-typedef std::unordered_map<uint, std::unordered_map<uint, std::pair<uint64_t, float>>> sd_table;
+typedef std::unordered_map<uint, std::unordered_map<uint, std::pair<uint64_t, double>>> sd_table;
 
 // Output current date and time in format specified with DMTTMFMT and DTTMSZ
 char* print_time (char *buff);
